Refactor file name Mask.ts to mask-controller.ts
