+++ src/main/ngapp/accio/src/app/editor/magic-wand.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { MaskController } from './Mask';
+import { MaskController } from './mask-controller';
 
 @Injectable({
   providedIn: 'root'
