+++ src/main/ngapp/accio/src/app/editor/mask-controller.ts
